requirements.txt
Django==1.8
MySQL-python==1.2.5
argparse==1.2.1
django-admin-exporter==0.2.1
unicodecsv==0.14.1
wsgiref==0.1.2
